feat(openssl): ability to use a pre-built openssl binary

+++ test/tests/01-package/run.sh
@@ -177,7 +177,7 @@ fi
 
 # kong binaries
 
-if [ "$SSL_PROVIDER" = "openssl" ]; then
+if [[ "$SSL_PROVIDER" = "openssl" ]] && [[ ! -z "${RESTY_OPENSSL_VERSION:-}" ]]; then
   docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "/usr/local/kong/bin/openssl version | grep -q ${RESTY_OPENSSL_VERSION}"
 fi
